sweeralert2

+++ public/javascripts/line_run_admin.js
@@ -123,15 +123,16 @@ var app = angular.module("RunAdmin", ['ngAnimate', 'ui.bootstrap', 'ui.bootstrap
                 showCancelButton: true,
                 confirmButtonText: "Yes, delete it!",
                 confirmButtonColor: "#ec6c62"
-            }, function () {
-
-                $http.delete("/api/runs/line/" + runIds).then(function (response) {
-                    console.log(response)
-                    updateRunList()
-                }, function (error) {
-                    console.log(error)
-                })
-            });
+            }).then((result) => {
+                if (result.value) {
+                    $http.delete("/api/runs/line/" + runIds).then(function (response) {
+                        console.log(response)
+                        updateRunList()
+                    }, function (error) {
+                        console.log(error)
+                    })
+                }
+            })
         }
 
         function updateRunList() {
@@ -149,9 +150,11 @@ var app = angular.module("RunAdmin", ['ngAnimate', 'ui.bootstrap', 'ui.bootstrap
                 showCancelButton: true,
                 confirmButtonText: "GO!",
                 confirmButtonColor: "#ec6c62"
-            }, function () {
-                $scope.go('/line/sign/' + runid + '/');
-            });
+            }).then((result) => {
+                if (result.value) {
+                    $scope.go('/line/sign/' + runid + '/');
+                }
+            })
         }
 
         $scope.go_judge = function (runid) {
@@ -162,9 +165,11 @@ var app = angular.module("RunAdmin", ['ngAnimate', 'ui.bootstrap', 'ui.bootstrap
                 showCancelButton: true,
                 confirmButtonText: "GO!",
                 confirmButtonColor: "#ec6c62"
-            }, function () {
-                $scope.go('/line/judge/' + runid + '/');
-            });
+            }).then((result) => {
+                if (result.value) {
+                    $scope.go('/line/judge/' + runid + '/');
+                }
+            })
         }
 
         $scope.go_approval = function (runid) {
@@ -175,9 +180,11 @@ var app = angular.module("RunAdmin", ['ngAnimate', 'ui.bootstrap', 'ui.bootstrap
                 showCancelButton: true,
                 confirmButtonText: "GO!",
                 confirmButtonColor: "#ec6c62"
-            }, function () {
-                $scope.go('/admin/approval/' + runid + '/');
-            });
+            }).then((result) => {
+                if (result.value) {
+                    $scope.go('/admin/approval/' + runid + '/');
+                }
+            })
         }
 
         $scope.go = function (path) {
